Refactor : generate 시 List<LottoNumber> 로 변경
