+++ src/test/java/lotto/vo/LottosTest.java
@@ -1,0 +1,15 @@
+package lotto.vo;
+
+import lotto.domain.Lotto;
+import lotto.service.NumberGenerateStrategy;
+import lotto.service.RandomLottoNumberGenerateStrategy;
+import org.junit.jupiter.api.Test;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class LottosTest {
+
+}
